# models/model_user_language_profile.go
/*
 * User Authentication API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type UserLanguageProfile struct {

	Language string `json:"language"`

	Actions []string `json:"actions"`
}