# models/model_user_sign_in_request.go
/*
 * User Authentication API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type UserSignInRequest struct {

	Username string `json:"username"`

	Password string `json:"password"`
}